requirements.txt
progbar-mushinako @ git+https://github.com/Mushinako/progbar.git
PyYAML
